% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clustering.R
\name{kmeans_clustering}
\alias{kmeans_clustering}
\title{Perform K-means clustering}
\usage{
kmeans_clustering(data, centers = 3, nstart = 10)
}
\arguments{
\item{data}{A numeric matrix or data frame.}

\item{centers}{The number of clusters.}

\item{nstart}{Number of initial random sets.}
}
\value{
A kmeans object.
}
\description{
This function applies K-means clustering to a dataset.
}
