% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot K-means clustering results}
\usage{
plot_clusters(data, kmeans_result)
}
\arguments{
\item{data}{A numeric matrix or data frame.}

\item{kmeans_result}{A kmeans object.}
}
\value{
A ggplot object.
}
\description{
This function visualizes K-means clusters.
}
